+++ client/src/gfx/effects.h
@@ -7,6 +7,6 @@ void fxExplosionBlaster (const vec pos,float pw);
 void fxBeamBlaster      (const vec pa,const vec pb, float beamSize, float damageMultiplier);
 void fxBlockBreak       (const vec pos, uchar b);
 void fxBlockMine        (const vec pos, int dmg, uchar b);
-void fxBleeding         (const vec pos, being victim, i16 dmg, u16 cause);
+void fxBleeding         (const vec pos, being victim, i16 dmg, u8 cause);
 void fxAnimalDiedPacket (const packet *p);
 void fxProjectileHit    (const packet *p);
